Batch process, preliminary conservation assessments

Combines the main of routines in rCAT to process multiple species for AOO, EOO etc.

Usage

batchCon(
  taxa,
  long,
  lat,
  project2gether = TRUE,
  cellsize = 2000,
  aooMin = FALSE,
  it = 1296,
  returnV = "S"
)

Arguments

taxa

field which defines a list of species or taxa

long

field which defines the longitude set of points

lat

field which defines the latitude set of points

project2gether

TRUE or FALSE, TRUE all data is projected together using the centre of all latitudes and longitudes. FALSE each species is projected separately. Default = TRUE

cellsize

cell length in metres used to for AOO projection N.B. IUCN recommend 2000 m (default 2000)

aooMin

calls the aooMin routines as well as simple aoo, be warned with lots of species and points this can take some time to run (default=FALSE)

it

if aooMin=TRUE this determines the number of iterations it will run to find aooMin (default=1296)

returnV

switches to return different sets of results:
S = simple returns a dataframe of results = (taxa ,Number of points,EOO in km2, Simple AOO in km2,Minimum AOO, EOO category, AOOcategory, Cellwidth, projection parameters)
SF = simple features dataframe will all results, taxa in taxon field, type=(EOO,AOO,points). NB all points will be projected together and aooMin is ignored

Details

This function expects a list of taxa and latitudes and longitudes.ie

species_w	85.388000	84.33100

species_w	-45.467000	88.41500

species_w	-34.339000	-149.52600

species_x	-29.620000	79.11900

species_x	33.409000	-33.94700

species_x	64.692000	-149.18900

species_y	2.308000	-140.21900

species_y	41.452000	-3.65600

species_y	-30.475000	-129.99600

etc

Note

Has a switch to either project all data as a whole or each taxa separately. We would suggest you use this switch if data is all from a similar area (i.e. all from one continent/country/region)

Author

Justin Moat. J.Moat@kew.org